eval/pyproject.toml
[project]
name = "passagent-eval"
version = "0.1.0"
description = "PassAgent evaluation suite"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"